requirements.txt
accelerate
bitsandbytes
datasets
numpy
sentencepiece
nltk
torch==2.4.0
transformers==4.49.0
openai
python-decouple
promptflow-tracing